Add Car demo (#181)
* Created Transform component

* Created GridRenderable component

* Created RenderSystem

* Re-implement car demo, both vanilla and ECS version

* Add transform and draw systems

* Adapt car demo to ECS systems

* Fix formatting

* Fix draw system argument name

+++ core/include/cubos/core/ecs/world_view.hpp
@@ -11,8 +11,8 @@ namespace cubos::core::ecs
     /// @tparam ComponentTypes The set of component types to be iterated.
     template <typename... ComponentTypes> struct WorldView
     {
-        World& world;              ///< The world being viewed.
-        std::vector<uint8_t> mask; /// The mask of the components to be iterated.
+        World& world;               ///< The world being viewed.
+        std::vector<uint32_t> mask; /// The mask of the components to be iterated.
 
         /// @param w The world to be viewed.
         WorldView(World& w);
@@ -20,14 +20,14 @@ namespace cubos::core::ecs
         /// Type of the iterator.
         struct Iterator
         {
-            World& world;                     ///< The world being viewed.
-            size_t current;                   ///< The current entity.
-            const std::vector<uint8_t>& mask; ///< The mask of the components to be iterated.
+            World& world;                      ///< The world being viewed.
+            size_t current;                    ///< The current entity.
+            const std::vector<uint32_t>& mask; ///< The mask of the components to be iterated.
 
             /// @param w The world to be viewed.
             /// @param index The current entity.
             /// @param m The mask of the components to be iterated.
-            Iterator(World& w, size_t index, const std::vector<uint8_t>& m);
+            Iterator(World& w, size_t index, const std::vector<uint32_t>& m);
 
             size_t operator*() const;
             bool operator==(const Iterator&) const;
@@ -54,7 +54,7 @@ namespace cubos::core::ecs
     }
 
     template <typename... ComponentTypes>
-    WorldView<ComponentTypes...>::Iterator::Iterator(World& w, size_t index, const std::vector<uint8_t>& m)
+    WorldView<ComponentTypes...>::Iterator::Iterator(World& w, size_t index, const std::vector<uint32_t>& m)
         : world(w), current(index), mask(m)
     {
     }
